+++ pkg/ansible/proxy/proxy.go
@@ -191,7 +191,7 @@ func addWatchToController(owner kubeconfig.NamespacedOwnerReference, cMap *contr
 	}
 	ownerGV, err := schema.ParseGroupVersion(owner.APIVersion)
 	if err != nil {
-		m := fmt.Sprintf("could not get broup version for: %v", owner)
+		m := fmt.Sprintf("could not get group version for: %v", owner)
 		log.Error(err, m)
 		return err
 	}
@@ -216,7 +216,7 @@ func addWatchToController(owner kubeconfig.NamespacedOwnerReference, cMap *contr
 	dependentPredicate := predicates.DependentPredicateFuncs()
 
 	// Add a watch to controller
-	if contents.WatchDependentResources {
+	if contents.WatchDependentResources && !gvkInList(resource.GroupVersionKind(), contents.Blacklist) {
 		// Store watch in map
 		// Use EnqueueRequestForOwner unless user has configured watching cluster scoped resources and we have to
 		switch {
@@ -248,8 +248,19 @@ func addWatchToController(owner kubeconfig.NamespacedOwnerReference, cMap *contr
 				return err
 			}
 		}
+	} else {
+		log.Info("Resource will not be watched/cached because it is blacklisted.", "GVK", resource.GroupVersionKind())
 	}
 	return nil
+}
+
+func gvkInList(current schema.GroupVersionKind, gvks []schema.GroupVersionKind) bool {
+	for _, gvk := range gvks {
+		if gvk == current {
+			return true
+		}
+	}
+	return false
 }
 
 func removeAuthorizationHeader(h http.Handler) http.Handler {