+++ pkg/ansible/run.go
@@ -125,6 +125,7 @@ func Run(flags *aoflags.AnsibleOperatorFlags) error {
 		}
 
 		cMap.Store(w.GroupVersionKind, &controllermap.Contents{Controller: *ctr,
+			Blacklist:                   w.Blacklist,
 			WatchDependentResources:     w.WatchDependentResources,
 			WatchClusterScopedResources: w.WatchClusterScopedResources,
 			OwnerWatchMap:               controllermap.NewWatchMap(),